Set acled parameters within layer configuration file
