+++ frontend/src/context/layers/point_data.ts
@@ -1,4 +1,3 @@
-import { camelCase } from 'lodash';
 import GeoJSON from 'geojson';
 import moment from 'moment';
 import type { LazyLoader } from './layer-data';
@@ -11,6 +10,7 @@ import { DEFAULT_DATE_FORMAT } from '../../utils/name-utils';
 import { fetchEWSData } from '../../utils/ews-utils';
 import { getAdminLevelDataLayerData } from './admin_level_data';
 import { fetchACLEDIncidents } from '../../utils/acled-utils';
+import { queryParamsToString } from '../../utils/url-utils';
 
 declare module 'geojson' {
   export const version: string;
@@ -22,24 +22,6 @@ declare module 'geojson' {
     callback?: Function,
   ): PointData[];
 }
-
-export const queryParamsToString = (queryParams?: {
-  [key: string]: string | { [key: string]: string };
-}): string =>
-  queryParams
-    ? Object.entries(queryParams)
-        .map(([key, value]) => {
-          if (key === 'filters') {
-            const filterValues = Object.entries(value)
-              .map(([filterKey, filterValue]) => `${filterKey}=${filterValue}`)
-              .join(',');
-
-            return `filters=${filterValues}`;
-          }
-          return `${camelCase(key)}=${value}`;
-        })
-        .join('&')
-    : '';
 
 export const fetchPointLayerData: LazyLoader<PointDataLayerProps> = () => async (
   {
@@ -68,7 +50,7 @@ export const fetchPointLayerData: LazyLoader<PointDataLayerProps> = () => async 
       case PointDataLoader.EWS:
         return fetchEWSData(dataUrl, date);
       case PointDataLoader.ACLED:
-        return fetchACLEDIncidents(dataUrl, date);
+        return fetchACLEDIncidents(dataUrl, date, additionalQueryParams);
       default:
         break;
     }
